fixed db, project of choice participants functions

+++ calendar_admin/bot_admin/models.py
@@ -6,7 +6,7 @@ logger = logging.getLogger(__name__)
 
 
 class User(models.Model):
-    user_tg_id = models.IntegerField(unique=True)
+    user_tg_id = models.BigIntegerField(unique=True)
     username = models.CharField(max_length=40, null=True)
     email = models.CharField(max_length=40, null=True)
     password_hash = models.CharField(max_length=150, null=True)
